Factor out division
